Add CPU/Display rate limiting.
* CPU.cs: Synchronize execution rate with a heartbeat timer.

* DebuggerPrompt.cs: Hack around DivByZero error when running in
  Xamarin Studio/Profiler console.

* EntryPoint.cs: Tweak to handle user interrupt with new CPU run loop.

* HighResolutionTimer.cs: Provide a cross-platform high-resolution
  timer for firing off real-time periodic events.

* ShugartController.cs: Try to avoid leaking a 25MB chunk of memory
at startup.

* Clock.cs:
* Z80System.cs: Clock counter from int to long.

* RasterOp.cs: Tiny documentation correction.

* PERQSystem.cs: Changes for modified CPU run loop.

* RawFloppyDisk.cs: Read PFD header if present in floppy images.

* AssemblyInfo.cs: Bump to v0.4.5 prior to interim release.

* Readme.txt: Documentation updates for interim release.

* SystemTimer.cs: Provide a heartbeat timer for CPU/Display rate
  limiting.

* PERQemu.csproj: Add new timer classes.

+++ PERQemu/IO/Z80/IOB/Clock.cs
@@ -102,7 +102,7 @@ namespace PERQemu.IO.Z80.IOB
             if (_enabled)
             {
                 int tickInterval = ((Z80System.Frequency / 60) / PERQCpu.IOFudge);
-                int elapsed = ((Z80System.Instance.Clocks() - _lastTick) / tickInterval);
+                int elapsed = ((int)(Z80System.Instance.Clocks() - _lastTick) / tickInterval);
 
                 fifo.Enqueue(Z80System.SOM);                    // SOM
                 fifo.Enqueue((byte)Z80toPERQMessage.ClockData); // Clock data message type
@@ -122,6 +122,6 @@ namespace PERQemu.IO.Z80.IOB
         private byte[] _messageData;
         private int _messageIndex;
         private bool _enabled = false;
-        private int _lastTick;
+        private long _lastTick;
     }
 }
